# Improve performance of denylist patterns