--- a/.editorconfig
+++ b/.editorconfig
@@ -9,6 +9,7 @@ end_of_line = lf
 insert_final_newline = true
 trim_trailing_whitespace = true
 indent_style = tab
+indent_size = 4
 
 [{.eslintrc,*.json,*.yml}]
 indent_style = space