This fixes an issue reported in a [support topic](https://wordpress.org/support/topic/long-url-with-many-query-parameters-slowing-loading/) where URLs with many query parameters can result in very slow response times by the service worker. Hat tip to @jeffposnick for pointing this out in https://github.com/GoogleChrome/workbox/issues/3077#issuecomment-1119033479.

The regular expressions in question are:

* `/\?(.+&)*wp_customize=/`
* `/\?(.+&)*customize_changeset_uuid=/`

Given this example URL: `https://example.com/?ab=4&v0=56&u0=degrees+Celsius+%28%C2%B0C%29&f0=C&v1=60&u1=degrees+Celsius+%28%C2%B0C%29&f1=C&v2=0.0005&u2=per+degree+Celsius+%28%2F%C2%B0C%29&f2=CdT&v3=100&u3=centimetres+%28cm%29&f3=cm&u4=centimetres+%28cm%29&f4=cm&v5=100&u5=sq+centimetres+%28cm%C2%B2%29&f5=sqcm&v6=100.4&u6=sq+centimetres+%28cm%C2%B2%29&f6=sqcm&v7=100&u7=cu+centimetres+%28cc%2C+cm%C2%B3%29&f7=cucm&v8=100.6&u8=cu+centimetres+%28cc%2C+cm%C2%B3%29&f8=cucm`

The patterns exhibit catastrophic backtracking:

[<img width="699" alt="image" src="https://user-images.githubusercontent.com/134745/167216822-1fb77e27-da8f-4125-b344-3f7364922362.png">](https://regexr.com/6l4qv)

This PR updates the patterns to be more efficient.

Old Pattern | Time
------------|--------------:
`^\/wp\-admin($\|\?.*\|\/.*)`         | 0.2ms
`[^\?]*.\.php($\|\?.*)`              | 0.7ms
`.*\?(.*&)?(wp_service_worker)=`    | 1.1ms
`.*\/wp\.serviceworker(\?.*)?$`     | 0.8ms
`[^\?]*\/feed\/(\w+\/)?$`           | 1.0ms
`\?(.+&)*wp_customize=`             | >250ms
`\?(.+&)*customize_changeset_uuid=` | >250ms
`^\/wp\-json\/.*`                   | 0.1ms

New Pattern | Time
-------|-------:
`^\/wp\-admin($\|\?\|\/)` | 0.0ms
`^[^\?]*?\.php($\|\?)` | 0.1ms
`\?(.*?&)?wp_service_worker=` | 0.1ms
`^[^\?]*?\/wp\.serviceworker(\?\|$)` | 0.1ms
`^[^\?]*?\/feed\/(\w+\/)?$` | 0.1ms
`\?(.*?&)?wp_customize=` | 0.1ms
`\?(.*?&)?customize_changeset_uuid=` | 0.1ms
`^\/wp\-json\/` | 0.0ms